server/requirements.txt
flask
keras
boto3
python-dotenv